Move tool configs into the ToolHandler + disallow overriding it in Agent.__call__
We remove overriding of tool_configs and tool_handler: The abstraction level here was a little odd - if we let you override it then you're dropping down to a very low level of tool execution.  Instead of dealing with AgentTool, you'd have to directly create the configs and the tool_handler (if you're injecting a new tool). Instead, we'll deprecate this in favor of some sort of hooking mechansim in the future, which would require an implementation of AgentTool directly - solving the problem of needing to generate the tool+config and the implementation(s) of overriden tools