chore: Remove agent.tool_config and update usages to use tool_specs
Agent.tool_config is a configuration object which serves as a wrapper to tool_specs and nothing else.  We actually don't use the toolChoice at all anywhere, and the `Tool` wrapper also was a container that served no purpose as everywhere we used the tools, we wanted the ToolSpec anyways.