feat: Enable hooks for MultiAgents
It's been a customer ask and we don't have a pressing need to keep it restricted.

The primary concern is that because agent's state is manipulated between invocations (state is reset) hooks designed for a single agent may not work for multi-agents. With documentation, we can guide folks to be aware of what happens rather than restricting it outright.