feat: Support adding exception notes for Python 3.10 (#1034)
When add_note is not available (3.10) enhance the default error message with the added notes.

In PR #290 we started using add_note to provide the bedrock model and region in exceptions to better clarify to customers what model & region were active.  The implementation used add_note which is only supported in 3.11+; however, we've had enough customers on 3.10 where they're not seeing the error message that it makes sense to add a shim to do something similar for 3.10.

---------

Co-authored-by: Mackenzie Zastrow <zastrowm@users.noreply.github.com>